(feat) newArticles loading at same time as the other sliders item
